Update PSMTabBarControl framework to fix icon problem
git-svn-id: https://nally.googlecode.com/svn/trunk@299 db928f3c-c32d-11dd-ba5c-d3b7c2b3b534

+++ YLController.mm
@@ -801,7 +801,7 @@
 //    [self refreshTabLabelNumber: _telnetView];
 }
 
-- (NSImage *) tabView: (NSTabView *)aTabView imageForTabViewItem: (NSTabViewItem *)tabViewItem offset: (NSSize *)offset styleMask: (unsigned int *)styleMask
+- (NSImage *) tabView: (NSTabView *)aTabView imageForTabViewItem: (NSTabViewItem *)tabViewItem offset: (NSSize *)offset styleMask: (NSUInteger *)styleMask
 {
     return nil;
 }
